Backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
